pyproject.toml
[project]
name = "news-of-the-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "feedparser>=6.0.12",
    "jupyterlab>=4.4.9",
    "langdetect>=1.0.9",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
